Move snd work on cripts for l3
cpppo_client4plc1 sends multiple cpppo ENIP write commands to
plc1 on the default ENIP port. The result is redirected
both to stdout and a dedicated .out log file, the errors are
logged into a dedicated .err file.

cpppo_plc1server listens for cpppo ENIP clients on the defualt
ENIP port. The result is redirected both to stdout and dedicated
a .out, the errors are logged into a dedicated .err file.